# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Fiber Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/api/feiras": {
            "post": {
                "description": "Feiras API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Feiras",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InsertUpdateFeiras"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/feiras/p/{page}": {
            "get": {
                "description": "Feiras API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Feiras",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/feiras/search": {
            "post": {
                "description": "Feiras API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Feiras",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchFeira"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/feiras/{id}": {
            "get": {
                "description": "Feiras API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Feiras",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id to search",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    }
                }
            },
            "put": {
                "description": "Feiras API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Feiras",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FeiraLivre"
                        }
                    },
                    {
                        "type": "string",
                        "description": "id to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Feira API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Feira",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/healthcheck": {
            "get": {
                "description": "HealthCheck API",
                "summary": "HealthCheck",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "models.FeiraLivre": {
            "type": "object",
            "properties": {
                "areap": {
                    "type": "integer"
                },
                "bairro": {
                    "type": "string"
                },
                "coddist": {
                    "type": "integer"
                },
                "codsubpref": {
                    "type": "integer"
                },
                "distrito": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lat": {
                    "type": "integer"
                },
                "logradouro": {
                    "type": "string"
                },
                "longi": {
                    "type": "integer"
                },
                "nome_feira": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "referencia": {
                    "type": "string"
                },
                "regiao5": {
                    "type": "string"
                },
                "regiao8": {
                    "type": "string"
                },
                "registro": {
                    "type": "string"
                },
                "setcens": {
                    "type": "integer"
                },
                "subprefe": {
                    "type": "string"
                }
            }
        },
        "models.InsertUpdateFeiras": {
            "type": "object",
            "properties": {
                "areap": {
                    "type": "integer"
                },
                "bairro": {
                    "type": "string"
                },
                "coddist": {
                    "type": "integer"
                },
                "codsubpref": {
                    "type": "integer"
                },
                "distrito": {
                    "type": "string"
                },
                "lat": {
                    "type": "integer"
                },
                "logradouro": {
                    "type": "string"
                },
                "longi": {
                    "type": "integer"
                },
                "nome_feira": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "referencia": {
                    "type": "string"
                },
                "regiao5": {
                    "type": "string"
                },
                "regiao8": {
                    "type": "string"
                },
                "registro": {
                    "type": "string"
                },
                "setcens": {
                    "type": "integer"
                },
                "subprefe": {
                    "type": "string"
                }
            }
        },
        "models.SearchFeira": {
            "type": "object",
            "properties": {
                "bairro": {
                    "type": "string"
                },
                "distrito": {
                    "type": "string"
                },
                "nomeFeira": {
                    "type": "string"
                },
                "pagina": {
                    "type": "string"
                },
                "regiao5": {
                    "type": "string"
                }
            }
        }
    }
}